refactor: move find_free_port to utils_network.py to resolve circular import
