+++ src/nanotron/utils.py
@@ -1,8 +1,6 @@
 import functools
 import inspect
 import os
-import random
-import socket
 from contextlib import ExitStack, contextmanager
 from typing import ContextManager, List, Optional
 
@@ -148,15 +146,3 @@ def tensor_from_untyped_storage(untyped_storage: torch.UntypedStorage, dtype: to
     tensor = torch.empty([], dtype=dtype, device=device)
     tensor.set_(source=untyped_storage)
     return tensor
-
-
-def find_free_port(min_port: int = 2000, max_port: int = 65000) -> int:
-    while True:
-        port = random.randint(min_port, max_port)
-        try:
-            with socket.socket() as sock:
-                sock.setsockopt(socket.SOL_SOCKET, socket.SO_REUSEADDR, 1)
-                sock.bind(("localhost", port))
-                return port
-        except OSError:
-            continue
